Implement Axis to Filter plugin

+++ UCR.Core/Utilities/Functions.cs
@@ -91,9 +91,9 @@ namespace HidWizards.UCR.Core.Utilities
             return value == short.MinValue ? short.MaxValue : Math.Abs(value);
         }
 
-        public static short GetRangeFromPercentage(short percentage)
+        public static short GetRangeFromPercentage(double percentage)
         {
-            if (percentage == -100) return Constants.AxisMinValue;
+            if ((short)percentage == -100) return Constants.AxisMinValue;
             return (short) (Constants.AxisMaxValue * (percentage / 100.0));
         }
     }
